refactor query assist input
remove unused feedback components
rename llm to query_assist
fix types
